[#7] Add doc URL into PositionInfo
This isn't available to the JSON/YAML parser, so it needs to be added
later when the instance is obtained for an overlay object.
