+++ json-overlay/src/main/java/com/reprezen/jsonoverlay/JsonLoader.java
@@ -28,7 +28,6 @@ import com.reprezen.jsonoverlay.parser.LocationRecorderJsonFactory;
 import com.reprezen.jsonoverlay.parser.LocationRecorderJsonParser;
 import com.reprezen.jsonoverlay.parser.LocationRecorderYamlFactory;
 import com.reprezen.jsonoverlay.parser.LocationRecorderYamlParser;
-import com.reprezen.jsonoverlay.parser.PositionInfo;
 
 public class JsonLoader {
